DicomSequence.split() added, bug fixes.
DicomSequence.split('attribute') now sets DicomSequence.ds to be a dict.
Keys are equal to 'attribute_name=attribute_value'. Each key is assigned
to a list containing a list of len 2 for every dicom file in the sequence.
First item in that list is the pydicom dataset for that file, and the
second item is the real path to that file.